+++ src/main/java/study3/domain/Results.java
@@ -8,7 +8,7 @@ import java.util.Map;
 public class Results {
 	
 	private Players players;
-	private LadderLine ladder;
+	private Ladder ladder;
 	private Rewards rewards;
 	
 	private static Map<Player, String> resultMap = new HashMap<>();
@@ -17,17 +17,17 @@ public class Results {
 		this.resultMap = resultMap;
 	}
 			
-	public Results(Players players, LadderLine ladder, Rewards rewards) {
+	public Results(Players players, Ladder ladder, Rewards rewards) {
 		this.players = players;
 		this.ladder = ladder;
 		this.rewards = rewards;
 	}
 
-	public static Results ofAll(Players players, LadderLine ladder, Rewards rewards) {
+	public static Results ofAll(Players players, Ladder ladder, Rewards rewards) {
 		return new Results(playGame(players, ladder, rewards));
 	}
 
-	private static Map<Player, String> playGame(Players players, LadderLine ladder, Rewards rewards) {
+	private static Map<Player, String> playGame(Players players, Ladder ladder, Rewards rewards) {
 		
 		int numberOfPlayer = players.getPlayersCount();
 		
@@ -42,11 +42,11 @@ public class Results {
 		return resultMap;
 	}
 
-	private static String findFinalPosition(Player siglePlayerName,List<Line> ladder, Rewards rewards) {
+	private static String findFinalPosition(Player siglePlayerName,List<LadderLine> ladder, Rewards rewards) {
 		
 		int position = Players.findPlayersIndex(siglePlayerName);
 		
-		for(Line line : ladder) {
+		for(LadderLine line : ladder) {
 			position = line.move(position);			
 		}		
 		return rewards.getReward(position);
