+++ src/main/java/study3/view/ResultView.java
@@ -10,8 +10,8 @@ import java.util.stream.Collectors;
 import study3.domain.Point;
 import study3.domain.Results;
 import study3.domain.Rewards;
+import study3.domain.Ladder;
 import study3.domain.LadderLine;
-import study3.domain.Line;
 import study3.domain.Player;
 import study3.domain.Players;
 
@@ -35,7 +35,7 @@ public class ResultView {
 		System.out.println();
 	}
 
-	public void viewRowAndColumn(LadderLine ladder) {
+	public void viewRowAndColumn(Ladder ladder) {
 
 		/* TODO : 레더 출력
 		 * for(int i=0; i<ladder.getLadder().size(); i++) {
@@ -48,7 +48,7 @@ public class ResultView {
 		});
 	}
 
-	private void choiceRowOrColum(Line line) {
+	private void choiceRowOrColum(LadderLine line) {
 		line.getLine().stream().forEach(point -> printBarAndColumn(getPointDirection(point)));
 		System.out.println();
 	}
