+++ src/test/java/nextstep/study3/LadderLineTest.java
@@ -3,7 +3,7 @@ package nextstep.study3;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import study3.domain.Line;
+import study3.domain.LadderLine;
 
 class LadderLineTest {
 		
@@ -11,14 +11,14 @@ class LadderLineTest {
 	@DisplayName("라인메소드 initLine 검증")
     public void init() {
         int sizeOfPerson = 5;
-        Line line = new Line(sizeOfPerson, false);
+        LadderLine line = new LadderLine(sizeOfPerson, false);
         System.out.println(line.getLine());
     }
 
     @Test
     @DisplayName("move 메소드 검증")
     public void move() {
-        Line line = new Line(2, false);
+        LadderLine line = new LadderLine(2, false);
         
         System.out.println("ladder result : " + line.move(0));
     }
